Permit disabling use of http_proxy
In the context of the HTCSS download plugin, it's not possible to
remove the http_proxy environment variable.  Add a second environment
variable to disable things.  Either of these are permitted:

- OSG_DISABLE_HTTP_PRXY
- OSDF_DISABLE_HTTP_PROXY

If the HTTP proxy is disabled, then fallback is also disabled.

Additionally, this permits it to accept OSDF_DISABLE_PROXY_FALLBACK
as an alias to OSG_DISABLE_PROXY_FALLBACK.